feat: flatpak, shares, antivirus and system updates
