+++ internal/commands/report/antivirus_linux.go
@@ -2,11 +2,9 @@
 
 package report
 
-import (
-	"errors"
-)
-
-// TODO LINUX
 func (r *Report) getAntivirusInfo(debug bool) error {
-	return errors.New("not implemented in Linux, yet")
+	r.Antivirus.Name = ""
+	r.Antivirus.IsActive = false
+	r.Antivirus.IsUpdated = false
+	return nil
 }
